--- packages/core/src/components/datetime-picker-rework/datetime-picker-rework.scss
@@ -1,57 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-@import 'common-variables';
-@import 'mixins/shadow-dom/component';
-
-:host {
-  @include ix-component;
-
-  display: block;
-  background-color: var(--theme-menu--background);
-  border-radius: 4px;
-  position: relative;
-  width: min-content;
-
-  .min-width {
-    width: min-content;
-  }
-
-  .no-padding {
-    padding: 0;
-  }
-
-  .flex {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-  }
-
-  @media (min-width: 576px) {
-    min-width: max-content;
-
-    .btn-select-date {
-      left: unset !important;
-    }
-
-  }
-
-  .individual {
-    box-shadow: none;
-    border: none;
-  }
-
-  .btn-select-date {
-    position: absolute;
-    bottom: $default-space;
-    right: $default-space;
-    left: $default-space;
-  }
-
-}
