+++ packages/core/src/components/datetime-picker/test/datetime-picker.ct.ts
@@ -9,10 +9,10 @@
 import { expect } from '@playwright/test';
 import { test } from '@utils/test';
 
-const DATE_TIME_PICKER_SELECTOR = 'ix-datetime-picker-rework';
+const DATE_TIME_PICKER_SELECTOR = 'ix-datetime-picker';
 
 test('renders', async ({ mount, page }) => {
-  await mount(`<ix-datetime-picker-rework></ix-datetime-picker-rework>`);
+  await mount(`<ix-datetime-picker></ix-datetime-picker>`);
   const datePicker = page.locator(DATE_TIME_PICKER_SELECTOR);
   await expect(datePicker).toHaveClass(/hydrated/);
 });
@@ -21,14 +21,14 @@ test.describe('datetime picker tests single', () => {
   test.beforeEach(async ({ mount }) => {
     await mount(
       `
-      <ix-datetime-picker-rework
+      <ix-datetime-picker
         range="false"
         from="1990/03/29"
         date-format="yyyy/LL/dd"
         time="09:10:12"
         time-format="HH:mm:ss"
         week-start-index="1"
-      ></ix-datetime-picker-rework>
+      ></ix-datetime-picker>
       `
     );
   });
