+++ packages/core/src/tests/datetime-picker-rework/basic/index.html
@@ -22,11 +22,11 @@ SPDX-License-Identifier: MIT
         margin-top: 2rem;
       "
     >
-      <ix-datetime-picker-rework
+      <ix-datetime-picker
         range="false"
         from="1990/03/29"
         time="09:10:12"
-      ></ix-datetime-picker-rework>
+      ></ix-datetime-picker>
     </div>
     <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
   </body>
